web/requirements.txt
# Django
asgiref==3.5.0
Django==4.0.4
sqlparse==0.4.2

# drf
djangorestframework==3.13.1

# Registration and auth
djoser==2.1.0

# For working with environments
django-environ==0.8.1

# Auto generate docs from drf api
drf-yasg==1.20.0